feat: taking out styles and using aliases
