+++ .eslintrc.cjs
@@ -26,6 +26,12 @@ module.exports = {
           ["@components", "./src/components"],
           ["@api", "./src/api"],
           ["@type", "./src/types"],
+          ["@pages", "./src/pages"],
+          ["@constants", "./src/constants"],
+          ["@hooks", "./src/hooks"],
+          ["@utils", "./src/utils"],
+          ["@ui", "./src/components/ui"],
+          ["@styles", "./src/styles"],
         ],
         extensions: [".ts", ".js", ".jsx", ".json"],
       },
