+++ src/App.tsx
@@ -2,9 +2,9 @@ import Layout from "@components/layout/Layout";
 import React from "react";
 import { Route, Routes } from "react-router-dom";
 import { ArtProvider } from "./context/ArtContext";
-import Detail from "./pages/Detail/Detail";
-import Favorites from "./pages/Favorites/Favorites";
-import Home from "./pages/Home";
+import Detail from "@pages/Detail/Detail";
+import Favorites from "@pages/Favorites/Favorites";
+import Home from "@pages/Home";
 
 const App: React.FC = () => {
   return (
