+++ src/components/layout/Layout.styles.ts
@@ -1,0 +1,11 @@
+import styled from "styled-components";
+
+export const PageWrapper = styled.div`
+  display: flex;
+  flex-direction: column;
+  min-height: 100vh;
+`;
+
+export const ContentWrapper = styled.main`
+  flex: 1;
+`;
