+++ src/components/layout/Layout.tsx
@@ -1,20 +1,10 @@
 import { Outlet } from "react-router-dom";
-import styled from "styled-components";
-import GlobalStyles from "../../styles/GlobalStyles";
 import ErrorBoundary from "../shared/ErrorBoundary";
 import Container from "./Container";
 import Footer from "./Footer/Footer";
 import Header from "./Header/Header";
-
-const PageWrapper = styled.div`
-  display: flex;
-  flex-direction: column;
-  min-height: 100vh;
-`;
-
-const ContentWrapper = styled.main`
-  flex: 1;
-`;
+import { ContentWrapper, PageWrapper } from "./Layout.styles";
+import GlobalStyles from "@styles/GlobalStyles";
 
 const Layout: React.FC = () => {
   return (
