+++ src/components/ui/Pagination/Pagination.tsx
@@ -1,39 +1,7 @@
 import { PaginationProps } from "@type/types";
 import React from "react";
-import styled from "styled-components";
-import { INITIAL_PAGE_INDEX, MAX_VISIBLE_PAGES } from "../../constants";
-
-const PaginationWrapper = styled.div`
-  display: flex;
-  justify-content: flex-end;
-  width: 100%;
-  margin: 20px 0;
-`;
-
-const PageButton = styled.button`
-  background: #fafafa;
-  color: #333;
-  border: none;
-  border-radius: 5px;
-  padding: 5px 10px;
-  margin: 0 5px;
-  cursor: pointer;
-
-  &:hover {
-    background: #ff7a00;
-    color: #fff;
-  }
-
-  &.active {
-    background: #ff7a00;
-    color: #fff;
-  }
-
-  &:disabled {
-    opacity: 0;
-    cursor: default;
-  }
-`;
+import { PageButton, PaginationWrapper } from "./Pagination.styles";
+import { MAX_VISIBLE_PAGES, INITIAL_PAGE_INDEX } from "@constants/constants";
 
 const Pagination: React.FC<PaginationProps> = ({
   currentPage,
