+++ src/components/ui/TitleSection/TitleSection.styles.ts
@@ -1,0 +1,18 @@
+import styled from "styled-components";
+
+export const Header = styled.div`
+  text-align: center;
+  margin-bottom: 30px;
+`;
+
+export const Subtitle = styled.div`
+  color: #e0a449;
+  font-size: 1rem;
+`;
+
+export const Title = styled.h1`
+  font-size: 32px;
+  color: #333;
+  margin: 0;
+  padding-bottom: 10px;
+`;
