+++ src/components/ui/TitleSection/TitleSection.tsx
@@ -1,23 +1,6 @@
 import { TitleSectionProps } from "@type/types";
 import React from "react";
-import styled from "styled-components";
-
-const Header = styled.div`
-  text-align: center;
-  margin-bottom: 30px;
-`;
-
-const Subtitle = styled.div`
-  color: #e0a449;
-  font-size: 1rem;
-`;
-
-const Title = styled.h1`
-  font-size: 32px;
-  color: #333;
-  margin: 0;
-  padding-bottom: 10px;
-`;
+import { Header, Subtitle, Title } from "./TitleSection.styles";
 
 const TitleSection: React.FC<TitleSectionProps> = ({ subtitle, title }) => {
   return (
