+++ src/pages/Detail/Detail.tsx
@@ -1,5 +1,5 @@
-import BookmarkButton from "@components/ui/BookmarkButton";
-import Spinner from "@components/ui/Spinner";
+import BookmarkButton from "@components/ui/BookmarkButton/BookmarkButton";
+import Spinner from "@components/ui/Spinner/Spinner";
 import { Artwork } from "@type/types";
 import React, { useEffect, useState } from "react";
 import { useParams } from "react-router-dom";
