+++ vite.config.ts
@@ -14,6 +14,12 @@ export default defineConfig({
       ),
       "@api": fileURLToPath(new URL("./src/api", import.meta.url)),
       "@type": fileURLToPath(new URL("./src/types", import.meta.url)),
+      "@pages": fileURLToPath(new URL("./src/pages", import.meta.url)),
+      "@constants": fileURLToPath(new URL("./src/constants", import.meta.url)),
+      "@hooks": fileURLToPath(new URL("./src/hooks", import.meta.url)),
+      "@utils": fileURLToPath(new URL("./src/utils", import.meta.url)),
+      "@styles": fileURLToPath(new URL("./src/styles", import.meta.url)),
+      "@ui": fileURLToPath(new URL("./src/components/ui", import.meta.url)),
     },
   },
 });
